Allow for Multiple Questions to be Answered
This commit meets the need for users to answer multiple questions by creating a column for each question response that is named based off the keyword used to trigger the question. It also updates the conditionals in the controller to correspond to the state requirements.

Finally there are some minor content updates and a short README update.